Refactor Screengrab to work with strategy objects
* Update to use the latest Android tools
* Create screenshot callback interface for async handling
