+++ screengrab/screengrab-lib/src/main/java/tools.fastlane.screengrab/ScreenshotCallback.java
@@ -1,0 +1,7 @@
+package tools.fastlane.screengrab;
+
+import android.graphics.Bitmap;
+
+public interface ScreenshotCallback {
+    void screenshotCaptured(Bitmap bitmap);
+}
