+++ screengrab/screengrab-lib/src/main/java/tools.fastlane.screengrab/ScreenshotStrategy.java
@@ -1,0 +1,5 @@
+package tools.fastlane.screengrab;
+
+public interface ScreenshotStrategy {
+    void takeScreenshot(ScreenshotCallback screenshotCallback);
+}
